# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SetPoint — Tennis Prematch API",
    "version": "1.0.0",
    "description": "Public REST API for SetPoint, a tennis prematch prediction app. Returns matches with Elo+Forme+Surface+H2H probabilities, multi-bookmaker odds (vig-removed), H2H history, Elo progression, and push/email value-bet alert subscriptions.\n\nBase URL: `https://setpoint.example`. All endpoints are JSON in / JSON out.",
    "contact": {
      "name": "SetPoint",
      "email": "contact@setpoint.example"
    },
    "license": {
      "name": "Proprietary",
      "url": "https://setpoint.example/license"
    }
  },
  "servers": [
    {
      "url": "https://setpoint.example",
      "description": "Production"
    },
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    }
  ],
...
  "components": {
    "schemas": {
...
      "EloHistoryPoint": {
        "type": "object",
        "required": ["date", "elo"],
        "properties": {
          "date": {
            "type": "string",
            "format": "date-time"
          },
          "elo": {
            "type": "integer"
          }
        }
      },
      "PlayerEloHistory": {
        "type": "object",
        "required": ["playerId", "currentElo", "history"],
        "properties": {
          "playerId": { "type": "string" },
          "currentElo": { "type": "integer" },
          "history": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/EloHistoryPoint" }
          }
        }
      },
      "EloHistoryResponse": {
        "type": "object",
        "required": ["matchId", "a", "b"],
        "properties": {
          "matchId": { "type": "string", "example": "m1" },
          "a": { "$ref": "#/components/schemas/PlayerEloHistory" },
          "b": { "$ref": "#/components/schemas/PlayerEloHistory" }
        }
      },
      "PushSubscription": {
        "type": "object",
        "required": ["endpoint", "keys"],
        "description": "Standard W3C PushSubscription object returned by `serviceWorkerRegistration.pushManager.subscribe()`.",
        "properties": {
          "endpoint": {
            "type": "string",
            "format": "uri",
            "description": "Browser-specific push service endpoint URL",
            "example": "https://fcm.googleapis.com/fcm/send/abc123"
          },
          "expirationTime": {
            "type": ["integer", "null"],
            "description": "Subscription expiration timestamp (ms since epoch), or null if it never expires"
          },
          "keys": {
            "type": "object",
            "required": ["p256dh", "auth"],
            "properties": {
              "p256dh": {
                "type": "string",
                "description": "ECDH P-256 public key, base64url"
              },
              "auth": {
                "type": "string",
                "description": "Authentication secret, base64url"
              }
            }
          }
        }
      },
      "PushSubscribeResponse": {
        "type": "object",
        "required": ["success", "count"],
        "properties": {
          "success": { "type": "boolean", "example": true },
          "count": {
            "type": "integer",
            "description": "Total number of registered subscriptions after this call",
            "example": 3
          }
        }
      },
...
      "EmailSubscription": {
        "type": "object",
        "required": ["email"],
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "description": "Email address to subscribe / unsubscribe",
            "example": "fan@setpoint.example"
          }
        }
      },
      "EmailSubscribeResponse": {
        "type": "object",
        "required": ["success", "count"],
        "properties": {
          "success": { "type": "boolean", "example": true },
          "count": {
            "type": "integer",
            "description": "Total number of subscribers after this call",
            "example": 12
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": {
            "type": "string",
            "description": "Human-readable error message",
            "example": "Invalid email format"
          }
        }
      }
    },
    "securitySchemes": {}
  },
  "paths": {
...
    "/api/tennis/elo-history": {
      "get": {
        "tags": ["tennis"],
        "summary": "Get Elo progression for both players of a match",
        "description": "Returns the 12-month Elo progression for both players of the requested match. Uses real Elo computed offline from Jeff Sackmann's tennis_atp / tennis_wta match CSVs (3-year window) for the 6 tracked players; falls back to a reverse-computed synthetic history for unknown player ids.",
        "operationId": "getEloHistory",
        "parameters": [
          {
            "name": "matchId",
            "in": "query",
            "required": true,
            "description": "Match identifier (one of `m1`, `m2`, `m3` in the mock dataset)",
            "schema": { "type": "string" },
            "example": "m1"
          }
        ],
        "responses": {
          "200": {
            "description": "Elo history for both players",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/EloHistoryResponse" },
                "examples": {
                  "default": {
                    "value": {
                      "matchId": "m1",
                      "a": {
                        "playerId": "sabalenka",
                        "currentElo": 2052,
                        "history": [
                          { "date": "2023-07-01T00:00:00Z", "elo": 1980 },
                          { "date": "2024-01-21T00:00:00Z", "elo": 2010 },
                          { "date": "2025-06-15T00:00:00Z", "elo": 2052 }
                        ]
                      },
                      "b": {
                        "playerId": "osaka",
                        "currentElo": 1759,
                        "history": [
                          { "date": "2023-07-01T00:00:00Z", "elo": 1820 },
                          { "date": "2024-01-21T00:00:00Z", "elo": 1780 },
                          { "date": "2025-06-15T00:00:00Z", "elo": 1759 }
                        ]
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing `matchId` query parameter",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "examples": {
                  "default": {
                    "value": { "error": "Missing matchId query parameter" }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error while computing the Elo history",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "examples": {
                  "default": {
                    "value": { "error": "Failed to compute Elo history" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/push/subscribe": {
      "post": {
        "tags": ["push"],
        "summary": "Register a Web Push subscription",
        "description": "Stores a W3C `PushSubscription` (returned by `serviceWorkerRegistration.pushManager.subscribe()`) in the in-memory subscription store. Deduplicates by `endpoint` — re-subscribing the same endpoint is idempotent. Returns 201 with the new total subscription count.",
        "operationId": "subscribePush",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/PushSubscription" },
              "examples": {
                "default": {
                  "value": {
                    "endpoint": "https://fcm.googleapis.com/fcm/send/eXaMpLe123",
                    "expirationTime": null,
                    "keys": {
                      "p256dh": "BOuY4f3k9...base64url",
                      "auth": "aBcDeFgH...base64url"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Subscription stored (or already existed)",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PushSubscribeResponse" },
                "examples": {
                  "default": {
                    "value": { "success": true, "count": 3 }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid subscription payload (missing or empty `endpoint`)",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "examples": {
                  "default": {
                    "value": { "error": "Invalid subscription payload" }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error while storing the subscription",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "examples": {
                  "default": {
                    "value": { "error": "Failed to store subscription" }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/email/subscribe": {
      "post": {
        "tags": ["email"],
        "summary": "Subscribe an email address to value-bet alerts",
        "description": "Validates the email format, normalizes it (trim + lowercase), dedupes by address (idempotent — re-subscribing the same address returns 201 without creating a duplicate), and stores it in the in-memory subscriber set. SMTP credentials are NEVER touched here — subscription only records the address.",
        "operationId": "subscribeEmail",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/EmailSubscription" },
              "examples": {
                "default": {
                  "value": { "email": "fan@setpoint.example" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Email subscribed (or already registered)",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/EmailSubscribeResponse" },
                "examples": {
                  "default": {
                    "value": { "success": true, "count": 12 }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid email format",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "examples": {
                  "default": {
                    "value": { "error": "Invalid email format" }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error while storing the subscription",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "examples": {
                  "default": {
                    "value": { "error": "Failed to store subscription" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/email/unsubscribe": {
      "post": {
        "tags": ["email"],
        "summary": "Unsubscribe an email address from value-bet alerts",
        "description": "Removes the email from the in-memory subscriber set. Returns 200 even if the address was not registered (idempotent) so the client can clear local state optimistically without worrying about a 404.",
        "operationId": "unsubscribeEmail",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/EmailSubscription" },
              "examples": {
                "default": {
                  "value": { "email": "fan@setpoint.example" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Email unsubscribed (or was not registered)",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/EmailSubscribeResponse" },
                "examples": {
                  "default": {
                    "value": { "success": true, "count": 11 }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid email format",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "examples": {
                  "default": {
                    "value": { "error": "Invalid email format" }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error while removing the subscription",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "examples": {
                  "default": {
                    "value": { "error": "Failed to remove subscription" }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}